+++ packages/sdk/CHANGELOG.md
@@ -9,9 +9,19 @@ Not every commit needs to result in a change to this file (e.g. docs and chore c
 commit that affects the code in a way that consumers might care about should include edits to the
 'Unreleased' section though. Breaking changes should be prefixed with `**BREAKING:**`.
 
-## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.55...HEAD)
+## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@1.0.0...HEAD)
 
 No changes.
+
+## [1.0.0](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.55...@airtable/blocks@1.0.0) - 2020-09-14
+
+-   **BREAKING:** The cell value format for lookup fields (`FieldType.MULTIPLE_LOOKUP_VALUES`) is
+    now `Array<{linkedRecordId: RecordId, value: CellValue}>`
+-   **BREAKING:** Remove `Record.primaryCellValue` and `Record.primaryCellValueAsString`, as well as
+    the `primaryCellValue` watchable key on Record. These APIs were deprecated in v0.0.45
+-   **BREAKING:** Remove legacy record creation format. All calls to `Table.createRecordAsync` /
+    `Table.createRecordsAsync` must define record field mappings under a `fields` key. These APIs
+    were deprecated in v0.0.41.
 
 ## [0.0.55](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.54...@airtable/blocks@0.0.55) - 2020-09-02